pyproject.toml
[tool.poetry]
name = "ebdms"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = " >=3.12,<4.0"
qrcode = {extras = ["pil"], version = "^8.2"}
django-crispy-forms = "^2.4"
crispy-bootstrap5 = "^2025.6"
django-reversion = "^6.1.0"
psycopg2 = "^2.9.11"
python-dotenv = "^1.2.1"
pydantic = "^2.12.5"
django-unfold = "^0.74.1"
django-revision = "^2.1.0"
django = "6.0"
django-jsonforms = "^1.1.2"
setuptools = "^80.9.0"
django-import-export = "^4.3.14"
openpyxl = "^3.1.5"
pandas = "^2.3.3"
django-otp = "^1.6.3"
requests = "^2.32.5"
django-rest-framework = "^0.1.0"
markdown = "^3.10"
django-filter = "^25.2"
django-storages = "^1.14.6"
boto3 = "^1.42.17"
django-minio-backend = "^4.5.0"
django-simple-history = "^3.11.0"


[tool.poetry.group.dev.dependencies]
ruff = "^0.14.10"
pylint = "^4.0.4"
bandit = "^1.9.2"
isort = "^7.0.0"
wat = "^0.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
